### 2025-10-27 #9315 @editorialbot
[PRE REVIEW]: Ellip: An Elliptic Integral Library for Rust
**Submitting author:** @p-sira (<a href="http://orcid.org/0000-0002-5636-8870">Sira Pornsiriprasert</a>)
**Repository:** https://github.com/p-sira/ellip
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.5.6
