### 2025-10-29 fafb850
PAPER: Format reference titles
