+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @book{abramowitz2013,
-  title = {Handbook of Mathematical Functions: With Formulas, Graphs, and Mathematical Tables},
+  title = {Handbook of {{Mathematical Functions}}: {{With Formulas}}, {{Graphs}}, and {{Mathematical Tables}}},
   shorttitle = {Handbook of Mathematical Functions},
   date = {2013},
   series = {Dover Books on Mathematics},
@@ -35,7 +35,7 @@
 }
 
 @article{bern2022,
-  title = {Scattering {{Amplitudes}}, the {{Tail Effect}}, and {{Conservative Binary Dynamics}} at \$\textbackslash mathcal\{\vphantom\}{{O}}\vphantom\{\}(\{\vphantom\}{{G}}\vphantom\{\}\textasciicircum\{4\})\$},
+  title = {Scattering {{Amplitudes}}, the {{Tail Effect}}, and {{Conservative Binary Dynamics}} at {{O}}({{G}}⁴)},
   author = {Bern, Zvi and Parra-Martinez, Julio and Roiban, Radu and Ruf, Michael S. and Shen, Chia-Hsien and Solon, Mikhail P. and Zeng, Mao},
   date = {2022-04-22},
   journaltitle = {Physical Review Letters},
@@ -70,7 +70,7 @@
 }
 
 @article{bulirsch1969,
-  title = {Numerical Calculation of Elliptic Integrals and Elliptic Functions. {{III}}},
+  title = {Numerical {{Calculation}} of {{Elliptic Integrals}} and {{Elliptic Functions}}. {{III}}},
   author = {Bulirsch, R.},
   date = {1969-08-01},
   journaltitle = {Numerische Mathematik},
@@ -90,7 +90,7 @@
 }
 
 @book{byrd1971,
-  title = {Handbook of elliptic integrals for engineers and scientists},
+  title = {Handbook of Elliptic Integrals for Engineers and Scientists},
   author = {Byrd, Paul F. and Friedman, Morris David and Byrd, Paul Francis and Friedman, Morris David},
   date = {1971},
   series = {Die Grundlehren der mathematischen Wissenschaften in Einzeldarstellungen mit besonderer Berücksichtigung der Anwendungsgebiete},
@@ -106,7 +106,7 @@
 }
 
 @article{caciagli2018,
-  title = {Exact Expression for the Magnetic Field of a Finite Cylinder with Arbitrary Uniform Magnetization},
+  title = {Exact {{Expression}} for the {{Magnetic Field}} of a {{Finite Cylinder}} with {{Arbitrary Uniform Magnetization}}},
   author = {Caciagli, Alessio and Baars, Roel J. and Philipse, Albert P. and Kuipers, Bonny W. M.},
   date = {2018-06-15},
   journaltitle = {Journal of Magnetism and Magnetic Materials},
@@ -134,7 +134,7 @@
 }
 
 @article{chandrupatla2010,
-  title = {{{THE PERIMETER OF AN ELLIPSE}}.},
+  title = {The {{Perimeter}} of an {{Ellipse}}},
   author = {Chandrupatla, Tirupathi R. and Osler, Thomas J.},
   date = {2010-12-01},
   journaltitle = {Mathematical Scientist},
@@ -154,7 +154,7 @@
 }
 
 @article{derby2010,
-  title = {Cylindrical Magnets and Ideal Solenoids},
+  title = {Cylindrical {{Magnets}} and {{Ideal Solenoids}}},
   author = {Derby, Norman and Olbert, Stanislaw},
   date = {2010-03-01},
   journaltitle = {American Journal of Physics},
@@ -176,13 +176,13 @@
[…]
