### 2025-10-29 1e1cb2a
PAPER: Format reference titles
