+++ paper/paper.bib
@@ -35,7 +35,7 @@
 }
 
 @article{bern2022,
-  title = {Scattering {{Amplitudes}}, the {{Tail Effect}}, and {{Conservative Binary Dynamics}} at {{O}}({{G}}⁴)},
+  title = {Scattering {{Amplitudes}}, the {{Tail Effect}}, and {{Conservative Binary Dynamics}} at {{O}}({{G}}\textasciicircum 4)},
   author = {Bern, Zvi and Parra-Martinez, Julio and Roiban, Radu and Ruf, Michael S. and Shen, Chia-Hsien and Solon, Mikhail P. and Zeng, Mao},
   date = {2022-04-22},
   journaltitle = {Physical Review Letters},
