### 2025-10-29 #9315 @p-sira
@danielskatz Thank you for your input. Table 5, Summary of Function Accuracy at 64-bit Precision, takes about 500 words, but it is essential to a mathematical software. It is mainly numbers, which I don't think should count toward the word count. Without Table 5, the word count should be about 1,100. Also, the Table 1 through 4 are mathematical definitions, taking about 280 words. These might be helpful for readers who are unfamiliar with the function names. Removing those together, the submission reaches just below 900 words. What are your recommendations?

Regarding the missing DOI that the bot suggested, it is not applicable to cite a book review for the book used.
