### 2025-11-03 #9315 @danielskatz
In case you have time to do this...
