### 2025-11-10 #9386 @editorialbot
[REVIEW]: Ellip: An Elliptic Integral Library for Rust
**Submitting author:** @p-sira (<a href="http://orcid.org/0000-0002-5636-8870">Sira Pornsiriprasert</a>)
**Repository:** https://github.com/p-sira/ellip
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.0
