### 2025-11-13 #9386 @joeybernard
The benchmark results should have machine details included (CPU family, speed, RAM, etc). Simply saying that it was 64-bits isn't enough to know what to expect when I run the tests on my machine.
