### 2025-11-17 #9386 @ariostas
@p-sira A few comments so far.

- It would be good to add arm64 testing on the CI. One of the tests fails for me on macOS arm64.
- If I run `cargo bench` I get similar numbers to the ones you report, but in µs, not ns.
- I think it would be nice to also update the versions in the `Cargo.toml` of the other two crates, since if you clone the repo it might be confusing to know which version you're actually using.

Also, just as a suggestion, since you have 3 crates in the repo, you might want to consider creating a [workspace](https://doc.rust-lang.org/book/ch14-03-cargo-workspaces.html) to keep things a bit more organized. But that's up to you.
