### 2025-11-19 060d088
DOC: Add clock speed to the test report
