### 2025-11-19 6669194
FMT: CI cargo fmt
