### 2025-11-19 b16f330
DOC: Format clock speed
