+++ examples/generate_test_summary.rs
@@ -5,7 +5,7 @@
 
 use ellip::*;
 use ellip_dev_utils::{
-    env::{format_clock_speed, get_env},
+    env::{format_cpu_with_clock_speed, get_env},
     get_summary_entry,
     test_report::generate_summary_table,
 };
@@ -17,8 +17,8 @@ fn main() {
 
     let summary_section = [
         &format!(
-            "Benchmark on {} @{} running {} rustc {} using ellip v{} at `f64` precision (ε=2.2204460492503131e-16).\n",
-            env.cpu, format_clock_speed(env.clock_speed), env.platform, env.rust_version, env.ellip_version
+            "Benchmark on {} running {} rustc {} using ellip v{} at `f64` precision (ε=2.2204460492503131e-16).\n",
+            format_cpu_with_clock_speed(&env.cpu, env.clock_speed), env.platform, env.rust_version, env.ellip_version
         ),
         "### Legendre's Elliptic Integrals",
         &generate_summary_table(&[
