### 2025-11-19 f26dfd6
PAPER: Add CPU details to the test results
