### 2025-11-20 2804014
CHG, CI: - The workflow is for testing only! - Use the 'workspace' project structure. - All the children packages now follow the version of ellip. - Use `cargo-release` for the workflow.
