+++ ellip-plot-graph/src/bin/ellippiinc_3d.rs
@@ -6,9 +6,9 @@
 use ellip::ellippiinc;
 use ellip_plot_graph::*;
 use plotly::{
+    Layout, Plot, Surface,
     common::{ColorScale, ColorScalePalette},
     layout::{Annotation, AspectRatio, Axis, Camera, CameraCenter, Eye, LayoutScene},
-    Layout, Plot, Surface,
 };
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
