+++ .github/workflows/publish.yml
@@ -17,6 +17,9 @@ jobs:
         with:
           fetch-depth: 0
           persist-credentials: true
+      - run: |
+          git config --global user.name 'github-actions'
+          git config --global user.email 'github-actions@github.com'
       - name: Extract version from tag
         id: version
         run: |
