+++ .github/workflows/publish.yml
@@ -25,6 +25,16 @@ jobs:
         run: |
           TAG="${GITHUB_REF#refs/tags/}"
           echo "version=${TAG#v}" >> $GITHUB_OUTPUT
+      - name: Cache cargo registry and build
+        uses: actions/cache@v4
+        with:
+          path: |
+            ~/.cargo/registry
+            ~/.cargo/git
+            target
+          key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
+          restore-keys: |
+            ${{ runner.os }}-cargo-
       - name: Setup Rust
         run: |
           rustup update stable
