+++ .github/workflows/publish.yml
@@ -5,7 +5,7 @@ on:
       - "v*"
 permissions:
   # contents: write
-  contents: read
+  contents: write
 jobs:
   publish:
     runs-on: ubuntu-latest
@@ -15,10 +15,7 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
         with:
-          fetch-depth: 0
-          persist-credentials: true
-      - name: Fix Git ownership in container
-        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
+          fetch-depth: 1
       - run: |
           git config --global user.name 'github-actions'
           git config --global user.email 'github-actions@github.com'
