+++ .github/workflows/publish.yml
@@ -16,6 +16,8 @@ jobs:
         uses: actions/checkout@v4
         with:
           fetch-depth: 1
+      - name: Fix Git ownership in container
+        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
       - run: |
           git config --global user.name 'github-actions'
           git config --global user.email 'github-actions@github.com'
