+++ .github/workflows/publish.yml
@@ -9,15 +9,11 @@ permissions:
 jobs:
   publish:
     runs-on: ubuntu-latest
-    container:
-      image: rust:latest
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
         with:
           fetch-depth: 1
-      - name: Fix Git ownership in container
-        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
       - run: |
           git config --global user.name 'github-actions'
           git config --global user.email 'github-actions@github.com'
@@ -42,7 +38,7 @@ jobs:
           cargo install cargo-release
       - name: Publish all
         # run: cargo release --workspace --no-tag "${{ steps.version.outputs.version }}" --execute
-        run: cargo release --workspace --no-tag --no-push "${{ steps.version.outputs.version }}"
+        run: cargo release --workspace --no-tag --no-push --allow-dirty "${{ steps.version.outputs.version }}"
         # env:
         #   CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
       - name: Archive workspace
