### 2025-11-20 7c30d80
CI: Auto generate CHANGELOG
