+++ .github/workflows/publish.yml
@@ -38,9 +38,15 @@ jobs:
           cargo install cargo-release
       - name: Publish all
         # run: cargo release --workspace --no-tag "${{ steps.version.outputs.version }}" --execute
-        run: cargo release --workspace --no-tag --no-push --allow-dirty "${{ steps.version.outputs.version }}"
+        run: cargo release --workspace --no-tag --no-push "${{ steps.version.outputs.version }}"
         # env:
         #   CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+      - uses: taiki-e/create-gh-release-action@v1
+        with:
+          # (Optional) Path to changelog.
+          changelog: CHANGELOG.md
+          # (Required) GitHub token for creating GitHub Releases.
+          token: ${{ secrets.GITHUB_TOKEN }}
       - name: Archive workspace
         run: |
           mkdir -p artifact
