+++ .github/workflows/publish.yml
@@ -4,7 +4,6 @@ on:
     tags:
       - "v*"
 permissions:
-  # contents: write
   contents: write
 jobs:
   publish:
@@ -37,22 +36,12 @@ jobs:
           rustup update stable
           cargo install cargo-release
       - name: Publish all
-        # run: cargo release --workspace --no-tag "${{ steps.version.outputs.version }}" --execute
-        run: cargo release --workspace --no-tag --no-push "${{ steps.version.outputs.version }}"
-        # env:
-        #   CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+        run: cargo release --workspace "${{ steps.version.outputs.version }}" --execute
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
       - uses: taiki-e/create-gh-release-action@v1
         with:
           # (Optional) Path to changelog.
           changelog: CHANGELOG.md
           # (Required) GitHub token for creating GitHub Releases.
           token: ${{ secrets.GITHUB_TOKEN }}
-      - name: Archive workspace
-        run: |
-          mkdir -p artifact
-          cp -r . artifact/
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: release-test-${{ steps.version.outputs.version }}
-          path: artifact
