### 2025-11-21 c705b20
CI: Remove prerelease version bump
