+++ .github/workflows/publish.yml
@@ -10,9 +10,6 @@ on:
           - patch
           - minor
           - major
-          - alpha
-          - beta
-          - rc
 permissions:
   contents: write
 jobs:
