### 2025-11-22 fffe6db
DOC: Improve testing documentation
