### 2025-11-22 a3893c1
CI: Add test on macOS and Windows in the CI
