### 2025-11-22 e66a23a
CI: Fix CI
