+++ .github/workflows/ellip-ci.yml
@@ -36,6 +36,7 @@ jobs:
     needs: test
     runs-on: ubuntu-latest
     steps:
+    - uses: actions/checkout@v4
     - name: Format code
       if: github.event_name == 'push'
       run: |
