### 2025-11-22 601db92
DOC: Update CHANGELOG
