+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 ## 0.5
 ### 0.5.7
 **Improvements**
+- Add reproducibility section in README.
 - Improve testing documentation.
 
 **Tests**
 - Add test on macOS and Windows in the CI.
+- Fix the elliprj test for macOS and Windows.
 
 **Others**
 - Use the 'workspace' project structure.
