### 2025-11-22 42d2bde
CI, DOC: Generate figures and test reports.
