### 2025-11-22 c749bea
DOC: Fix the tables in README
