+++ tests/README.md
@@ -1,140 +1,164 @@
 # Testing
-This report presents the accuracy of the ellip crate using **symmetric relative error**, defined as
 
-![Symmetric relative error](https://github.com/p-sira/ellip/blob/main/examples/symmetric_error.svg?raw=true)
+This report denotes how to reproduce the tests and presents the accuracy of the ellip crate using **symmetric relative error**, defined as
 
-Errors are expressed in units of machine epsilon (ε). The test data spans the domain of each function up to **μ** to avoid approaching the function's limit. The reference values are computed using [**Wolfram Engine**](https://www.wolfram.com/engine/). You can find the scripts in the directory [tests/wolfram/](https://github.com/p-sira/ellip/blob/main/tests/wolfram/). 
-This report is generated on AMD Ryzen 5 4600H with Radeon Graphics @3.0 GHz running x86_64-unknown-linux-gnu rustc 1.90.0 using ellip v0.5.6 at `f64` precision (ε=2.2204460492503131e-16).
+![Symmetric relative error](https://github.com/p-sira/ellip/blob/main/examples/symmetric_error_on_white.svg?raw=true)
 
-## Legendre's Complete Elliptic Integrals
+Errors are expressed in units of machine epsilon (ε). This report is generated using `cargo run --example generate_error_report`.
 
-| Function        | Mean (ε) | Median (ε) | P99 (ε) | Max (ε) | Variance (ε²) | μ (ε²) |
-|-----------------|----------|------------|---------|---------|---------------|--------|
-| ellipk          | 0.27     | 0.00       | 1.51    | 20.95   | 0.81          | 1      |
-| ellipk (Neg m)  | 0.54     | 0.00       | 1.55    | 108.14  | 15.93         | 1      |
-| ellipe          | 0.30     | 0.00       | 1.82    | 3.00    | 0.18          | 1      |
-| ellipe (Neg m)  | 0.42     | 0.51       | 1.64    | 1.95    | 0.20          | 1      |
-| ellippi         | 0.41     | 0.00       | 1.65    | 20.95   | 0.60          | 1      |
-| ellippi (Neg m) | 0.43     | 0.51       | 1.82    | 2.96    | 0.22          | 1      |
-| ellippi (p.v.)  | 0.17     | 0.00       | 2.11    | 36.35   | 2.56          | 50     |
-| ellipd          | 0.44     | 0.00       | 1.96    | 2.43    | 0.32          | 1      |
-| ellipd (Neg m)  | 0.10     | 0.00       | 1.87    | 2.64    | 0.16          | 1      |
+## Test Datasets
 
-## Legendre's Incomplete Elliptic Integrals
+Ellip uses three sources of reference values:
+1. **Wolfram Engine** for accuracy reporting and unit tests
+2. **Boost.Math test data** in unit tests
+3. **Original literature** in the unit tests for Bulirsh's integrals
 
-| Function                    | Mean (ε) | Median (ε) | P99 (ε) | Max (ε) | Variance (ε²) | μ (ε²) |
-|-----------------------------|----------|------------|---------|---------|---------------|--------|
-| ellipf                      | 0.33     | 0.00       | 2.39    | 7.47    | 0.42          | 1      |
-| ellipf (Neg m)              | 0.28     | 0.00       | 1.95    | 3.00    | 0.26          | 1      |
-| ellipeinc                   | 0.51     | 0.00       | 7.50    | 24.66   | 2.69          | 1      |
[…]
