### 2025-11-22 45cfe40
Revert "Merge branch 'main' into dev"
This reverts commit 882b12bad882e666c4261b7996bff47648665146, reversing
changes made to c749bea1ce283a376ff3f234d0b811532eecf8a4.
