### 2025-11-22 ff27654
CI, DOC: Generate figures and test reports.
