### 2025-11-22 feda210
PAPER: Add CPU details
