### 2025-11-25 #9386 @ariostas
@p-sira the workspace addition looks great!

Minor fix for references: You should change `zotero-item-1032` in the bib file to `@pornsiriprasert2025`.

The only part that I see missing is this.
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I didn't find this anywhere (hopefully I didn't miss it). A section in the readme is probably fine, or you can add `CONTRIBUTING.md` and `CODE_OF_CONDUCT.md` files to your repo and GitHub will add tabs/links in the main page of your repo.
