### 2025-11-26 9bbfef1
PAPER: Fix citation key
