+++ paper/paper.bib
@@ -291,7 +291,7 @@
   year = {2020}
 }
 
-@online{zotero-item-1032,
+@online{pornsiriprasert2025,
   title = {{{EllipPy Documentation}} — {{EllipPy}} 0.5.6},
   author = {Pornsiriprasert, Sira},
   date = {2025-10-17},
