### 2025-11-30 3048b17
DOC: Create CONTRIBUTING.md
