### 2025-11-30 eed7382
PAPER: Remove original date in plotly reference
