+++ paper/paper.bib
@@ -227,7 +227,6 @@
   title = {Plotly.Rs},
   author = {{Plotly Inc.}},
   date = {2025-01-03},
-  origdate = {2020-01-26T10:57:20Z},
   url = {https://github.com/plotly/plotly.rs},
   urldate = {2025-10-13},
   abstract = {Plotly for Rust},
