### 2025-11-30 #9386 @p-sira
@ariostas I've fixed the references and added `CONTRIBUTING` and `CODE_OF_CONDUCT`.
