### 2025-12-07 d1f5491
PAPER: Moved the figures to the summary section
