### 2025-12-07 40bf8fa
PAPER: Formatting
