+++ paper/paper.md
@@ -133,6 +133,8 @@ Table 3: Carlson's Symmetric Elliptic Integral Functions of the Module `carlson`
 | Computation using Carlson's duplication theorem [@carlson2025].                                                                                                                                                                     |
 +======================================================+==============================================================================================================================================================================+
 
+<br>
+
 Table 4: Miscellaneous Functions Related to Elliptic Integrals of the Module `misc`
 
 +------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
