### 2025-12-07 e5609cd
PAPER: Formatting
