+++ paper/paper.md
@@ -48,6 +48,10 @@ where $P$ is a cubic or quartic polynomial in $t$ [@byrd1971].
 
 Ellip consists of four modules: legendre, bulirsch, carlson, and misc. The functions are implemented exclusively in Rust, accept generic real numbers provided by the num-traits's Float [@stone2024], and operate entirely on the stack. The functions are outlined in Tables 1, 2, 3, and 4, respectively.
 
+Performance is optimized by deferring input validation, i.e., assuming inputs are valid and raising errors upon completing the routine. Ellip-Rayon was released as a companion library for parallelizing large inputs. Python is supported via the EllipPy library, using PyO3 [@pyo3projectandcontributors2025] for Rust-Python binding. The documentation covers all public functions, encompassing their mathematical definitions, domains, graphical representations, special cases, and related functions.
+
+The compilation of Ellip is controlled by feature flags. The `no-std` flag enables Ellip to compile in no-std environments. The `unstable` flag exposes internal functions for advanced users, bypassing input validations at the cost of safety guarantees. Lastly, the `test_force_fail` is used for code coverage, where some conditions are unreachable as remnants of defensive programming.
+
 Table 1: Legendre's Elliptic Integral Functions of the Module `legendre`
 
 +------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
@@ -133,8 +137,6 @@ Table 3: Carlson's Symmetric Elliptic Integral Functions of the Module `carlson`
 | Computation using Carlson's duplication theorem [@carlson2025].                                                                                                                                                                     |
 +======================================================+==============================================================================================================================================================================+
 
-<br>
-
 Table 4: Miscellaneous Functions Related to Elliptic Integrals of the Module `misc`
 
 +------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
@@ -147,10 +149,6 @@ Table 4: Miscellaneous Functions Related to Elliptic Integrals of the Module `mi
 +======================================================+==============================================================================================================================================================================+
 | Computation using relations to Carlson's symmetric integrals.                                                                                                                                                                       |
 +======================================================+==============================================================================================================================================================================+
-
-Performance is optimized by deferring input validation, i.e., assuming inputs are valid and raising errors upon completing the routine. Ellip-Rayon was released as a companion library for parallelizing large inputs. Python is supported via the EllipPy library, using PyO3 [@pyo3projectandcontributors2025] for Rust-Python binding. The documentation covers all public functions, encompassing their mathematical definitions, domains, graphical representations, special cases, and related functions.
-
-The compilation of Ellip is controlled by feature flags. The `no-std` flag enables Ellip to compile in no-std environments. The `unstable` flag exposes internal functions for advanced users, bypassing input validations at the cost of safety guarantees. Lastly, the `test_force_fail` is used for code coverage, where some conditions are unreachable as remnants of defensive programming.
 
 # Results
 
