### 2025-12-07 f134e09
PAPER: Formatting
