### 2025-12-07 4d15296
PAPER: Formatting
