### 2025-12-07 5bb2e16
PAPER: Formatting
