+++ paper/paper.md
@@ -203,6 +203,7 @@ Table 5: Summary of Function Accuracy at 64-bit Precision.
 +---------------------+----------------+-------------+-------------------+
 | el3                 | $0.00$         | $53.21$     | $17.51$           |
 +---------------------+----------------+-------------+-------------------+
++---------------------+----------------+-------------+-------------------+
 | **Module `carlson`**                                                   |
 +---------------------+----------------+-------------+-------------------+
 | elliprf             | $0.00$         | $1.57$      | $0.19$            |
