### 2025-12-07 #9386 @p-sira
I've moved the graph figure to the summary section and rearranged the paragraphs to adjust the formatting a little bit.
