### 2025-12-15 660c396
Small formatting edits
