### 2025-12-19 3c758ca
DOC: Update CHANGELOG
