### 2025-12-19 a1aa3f5
PAPER: Update the version number
