### 2025-12-19 18c1f55
CI, DOC: Generate figures and test reports.
