### 2025-12-19 #9386 @RichardLitt
Ping @danielskatz
