### 2025-12-20 d265f34
BUG: Fix integer overflow in `ellipk` and `ellipe`.
