### 2025-12-20 4bd0ed2
CI, DOC: Generate figures and test reports.
