### 2025-12-29 64c0ed2
CHG: Deprecate no_std flag
