+++ src/carlson/elliprd.rs
@@ -197,7 +197,7 @@ const N_MAX_ITERATIONS: usize = 50;
 const N_MAX_ITERATIONS: usize = 1;
 
 #[cfg(not(feature = "test_force_fail"))]
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(test)]
 mod tests {
     use super::*;
     use crate::compare_test_data_boost;
