+++ src/legendre/ellipd.rs
@@ -90,7 +90,7 @@ pub fn ellipd<T: Float>(m: T) -> Result<T, StrErr> {
 }
 
 #[cfg(not(feature = "test_force_fail"))]
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(test)]
 mod tests {
     use super::*;
     use crate::{compare_test_data_boost, compare_test_data_wolfram};
