+++ src/util.rs
@@ -24,7 +24,7 @@ pub fn assert_close<T: Float>(actual: T, expected: T, rtol: T) {
 }
 
 #[cfg(not(feature = "test_force_fail"))]
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(test)]
 mod tests {
     use super::*;
 
