+++ src/legendre/ellipe.rs
@@ -290,7 +290,7 @@ fn ellipe_precise<T: Float>(m: T) -> Result<T, StrErr> {
 }
 
 #[cfg(not(feature = "test_force_fail"))]
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(test)]
 mod tests {
     use core::f64;
 
