+++ src/bulirsch/el.rs
@@ -630,7 +630,7 @@ const N_MAX_ITERATIONS: usize = 10;
 const N_MAX_ITERATIONS: usize = 1;
 
 #[cfg(not(feature = "test_force_fail"))]
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(test)]
 mod tests {
     use super::*;
     use crate::{assert_close, compare_test_data_wolfram};
