### 2025-12-29 d927366
CI: Remove no_std test
