### 2025-12-29 e7eece6
DOC, REL: Update CHANGELOG, Push version
