### 2025-12-29 2dfa8ff
CI, DOC: Generate figures and test reports.
