### 2026-01-18 #9386 @tarleb
My apologies, life interfered and I didn't manage to do it until now. The update goes life once this PR is merged: https://github.com/xuanxu/publishing-artifacts-action/pull/4
