### 2026-01-23 #9386 @tarleb
Oh, that's not good. This appears to be about MathML, but I'm not sure what the problem might be.
