### 2026-01-24 #9386 @tarleb
@xuanxu could this be a problem with the validator? My interpretation of the errors is that the validator expects `msup` to contain only a single child element, but there are multiple. I don't see anything in the [MathML specs](https://w3c.github.io/mathml-core/#subscripts-and-superscripts-msub-msup-msubsup) that would disallow multiple children in `msup`.
