### 2026-01-24 #9386 @xuanxu
The validator considers that the [generated JATS file](https://github.com/openjournals/joss-papers/blob/joss.09386/joss.09386/paper.jats/10.21105.joss.09386.jats) is invalid, do you spot any problem in it?. I'm not sure about the precise origin of the error, I'll try to verify the cause.
