### 2026-01-24 #9386 @tarleb
The validator at https://jats4r-validator.niso.org/ only complains about accessibility issues, but the JATS itself is valid according to that site.
