### 2026-01-24 #9386 @xuanxu
It looks like the file is invalid JATS v1.3 but valid v1.2.
I've updated our pipeline to accept the metadata file if that happens, @arfon / @tarleb WDYT?

Recommend acceptance can be retried now.
