### 2026-01-25 #9386 @tarleb
Oh, how interesting! I hadn't expected this kind of changes in JATS v1.3. Our current template specifies JATS 1.2 as DTD, so we should be ok for now.

I'll try to create a minimal example and open a pandoc issue.
