### 2026-01-31 #9386 @danielskatz
👋 @p-sira - As track editor, I suggest the changes in https://github.com/p-sira/ellip/pull/102 - please merge them, or let me know what you disagree with.

I note that you don't have an acknowledgments section. Is this what you intended?
