### 2026-02-01 caa1243
PAPER: Add acknowledgment section
