### 2026-02-01 #9386 @p-sira
@danielskatz Thank you for pointing this out. I'm happy to add the acknowledgment section. Also, I've commented on your PR in https://github.com/p-sira/ellip/pull/102.
