### 2026-02-02 #9386 @p-sira
@danielskatz I've merged the bib case change (https://github.com/p-sira/ellip/pull/102).
