### 2026-02-03 #9386 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
