### 2026-02-03 e4f2202
Update paper.bib
