+++ paper/paper.bib
@@ -238,7 +238,7 @@
 
 @software{pyo3projectandcontributors2025,
   title = {{PyO3}},
-  author = {{PyO3} Project and Contributors},
+  author = {{PyO3 Project and Contributors}},
   date = {2025-08-29},
   url = {https://github.com/PyO3/pyo3},
   urldate = {2025-10-14},
