### 2026-02-03 #9386 @danielskatz
@p-sira - Sorry, one more small change is needed - https://github.com/p-sira/ellip/pull/104 - this fixes an error I introduced.
