### 2026-02-03 #9386 @danielskatz
Congratulations to @p-sira (Sira Pornsiriprasert) on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you
